Changed up the FS on the project, little better organization this way. Treks, TrekCard, Trek class, and main are separate files.
